Changing build system to using CMake and GNU toolchain.
This is by no means done. Heck, I'm not even sure if the code does anything in this state, but I just got it to compile (FINALLY).
I will clean this up and start building the dependent libraries independently as far as possible. More to come.
In the meantime, and this is probably more of a note to self, this is how you build the project:
- Make a build folder in root of project: mkdir build
- Navigate into the build folder: cd build
- Call cmake and hand it the toolchain file: cmake -DCMAKE_TOOLCHAIN_FILE=~/Projects/SAMD21/toolchain_samd21g18a.cmake ..
- Now, build the project by calling: make

The project uses arm-none-eabi-gcc which you can install using your favorite package manager.