Adding Reworked Version of Shared RegisterGuard
I am not sure if I like this yet. At this point I am still just playing
around with different ideas. This still does not seem very bullet proof,
and it requires the user to be consistent with setting IsShared to true
for shared registers (wrongly setting this could be hard to track down,
as it only influences simulation).