Adding an "Allocator" Class for My DataStructures
Basically, the idea is to use static memory allocation and add Construct
and Destroy capabilities. This prevents us from having to have a default
constructor for Objects (as I previously replaced destroyed objects with
a default constructed object), and generally cleans up memory
management.

Unfortunately, I cannot use the standard library data structures, as the
std::allocator does not support reallocate functionality. Therefore, I
could not grow an existing vector, but would have to continuously
allocate larger memory blocks and copy over data. This is not feasible
on a system as memory capped as this one.