Picking this project back up
Messy commit as I was trying to figure out where I left off.
One of the big problems is the generlization of the HAL to support
implementations for both simulated hardware and SAMD HW. I feel like I
have been way overthinking this in search of a neat solution. Decided to
just add a virtual base class that describes the peripheral interface
and a definitions file with aliases that are replaced in a simulated
environment. This is by no means a great solution as it adds more
virtual call overhead. I may revisit this in the future, but for now
this will get me unblocked.

I have also removed the awkward CONFIG templates that weren't really
working anyways. Peripherals can be configured via the menuconfig tool
that generates a config.h.