Further cleanup of clock peripherals
I tried to use CRTP for all peripherals, but I don't think it is a great
solution. Clocks for instance, will be used by most other peripherals
(and separation is painful as a lot of operations have to be coordinated
with clock setup) and CRTP would require templatized versions of all
clock users (which forces me to  effectively put all the implementation
in header files). Although possible (and slightly more efficient), I
don't think its worth it when taking readability into consideration.